email the rse team when a request comes in
