+++ institution/fixtures/institutions.json
@@ -11,6 +11,7 @@
       "needs_legacy_inst_id": false,	
       "separate_allocation_requests": true,
       "allows_rse_requests": true,
+      "rse_notify_email": "sa2c-support@swansea.ac.uk",
       "local_repository_name": "Cronfa",
       "local_repository_domain": "cronfa.swansea.ac.uk",
       "funding_database_name": "AMS",
@@ -29,6 +30,7 @@
       "needs_legacy_inst_id": false,	
       "separate_allocation_requests": false,	
       "allows_rse_requests": false,
+      "rse_notify_email": null,
       "local_repository_name": "",
       "local_repository_domain": "",
       "funding_database_name": "",
@@ -47,6 +49,7 @@
       "needs_legacy_inst_id": true,	
       "separate_allocation_requests": false,	
       "allows_rse_requests": false,
+      "rse_notify_email": null,
       "local_repository_name": "",
       "local_repository_domain": "",
       "funding_database_name": "",
@@ -65,6 +68,7 @@
       "needs_legacy_inst_id": false,	
       "separate_allocation_requests": false,	
       "allows_rse_requests": false,
+      "rse_notify_email": null,
       "local_repository_name": "",
       "local_repository_domain": "",
       "funding_database_name": "",
