+++ openldap/api/project_api.py
@@ -5,12 +5,12 @@ from django.conf import settings
 from django.utils.translation import gettext_lazy as _
 from django_rq import job
 
+from common.util import email_user
 from openldap.schemas.project.activate_project import activate_project_json
 from openldap.schemas.project.create_project import create_project_json
 from openldap.schemas.project.get_project import get_project_json
 from openldap.schemas.project.list_projects import list_projects_json
 from openldap.util import decode_response
-from openldap.util import email_user
 from openldap.util import raise_for_data_error
 from openldap.util import verify_payload_data
